- DEVSU-2301 - Rewrite Reports Table as a separate component for future reusing (when creating new filtered reports views) - Revised padding and size of collapse button for the Reports section on the sidebar - Remove PatientsView component and all its subsequent routing
