+++ app/components/ReportsTable/index.scss
@@ -1,0 +1,4 @@
+.reports-table__container {
+  flex: 1;
+  height: 100%;
+}
