+++ app/views/MainView/index.tsx
@@ -36,7 +36,6 @@ const CondensedPrintView = (props) => <PrintView {...props} printVersion="conden
 const GermlineView = lazy(() => import('../GermlineView'));
 const AdminView = lazy(() => import('../AdminView'));
 const LinkOutView = lazy(() => import('../LinkOutView'));
-const PatientsView = lazy(() => import('../PatientsView'));
 const TemplateView = lazy(() => import('../TemplateView'));
 const ProjectsView = lazy(() => import('../ProjectsView'));
 
@@ -182,7 +181,6 @@ const Main = (): JSX.Element => {
                   <AuthenticatedRoute component={TermsView} path="/terms" />
                   <AuthenticatedRoute component={ReportsView} path="/reports" />
                   <AuthenticatedRoute component={MyReportsView} path="/my-reports" />
-                  <AuthenticatedRoute exact component={PatientsView} path="/reports/patients/:patientId" />
                   <Redirect exact from="/report/:ident/(genomic|probe)/summary" to="/report/:ident/summary" />
                   <AuthenticatedRoute component={ReportView} path="/report/:ident" />
                   <AuthenticatedRoute component={PrintView} path="/print/:ident" showNav={false} onToggleNav={setIsNavVisible} />
