--- app/views/MyReportsView/index.scss
@@ -1,4 +1,0 @@
-.reports-table__container {
-  flex: 1;
-  height: 100%;
-}
