+++ app/views/ReportsView/index.scss
@@ -1,4 +1,1 @@
-.reports-table__container {
-  flex: 1;
-  height: 100%;
-}
+
